Add test stub generator API
